% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EHR.R
\docType{data}
\name{EHR}
\alias{EHR}
\title{Description of the EHR dataset}
\format{
A data frame with \code{3000} records and \code{3} variables:
\describe{
\item{link_id}{Unique patient identifyer}
\item{reg}{Registered long-term condition}
\item{aos}{Age at onset of the registered long-term condition}

}
}
\source{
Toy dataset
}
\description{
This is a toy dataset to illustrate the use of the the MSCA library
}
\examples{
# Load the dataset
data(EHR)

# Display the first few rows
head(EHR)
}
