% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fast_jack_upper.R
\name{fast_jaccard_dist}
\alias{fast_jaccard_dist}
\title{Compute upper triangle Jaccard distance}
\usage{
fast_jaccard_dist(mat, as.dist = FALSE)
}
\arguments{
\item{mat}{A numeric binary matrix (0/1/NA)}

\item{as.dist}{Logical. If TRUE, return a dist object; otherwise an upper triangular matrix.}
}
\value{
A matrix or dist object of Jaccard dissimilarities (upper triangle)
}
\description{
Compute upper triangle Jaccard distance
}
