\name{MSCA}
\alias{MSCA-package}
\docType{package}
\title{Clustering of Multiple Censored Time-to-Event Endpoints}
\description{The MSCA library provides basic tools for computing clusters
    of instances described by multiple time-to-event censored endpoints.
    From long format datasets, where one instance is described by one or
    more record of events, a procedure is used to compute states matrices.
    Then, from states matrices, a procedure provides optimised computation
    of the Jacard distance between instances. The library is currently in
    development and more options and tools allowing graphical
    representation of typologies are expected.}
\details{
  This package is under active development
}
\author{

Marc Delord

\strong{Reference:} Delord M, Douiri A (2025) "Multiple states clustering analysis (MSCA),
  an unsupervised approach to multiple time-to-event electronic health records
  applied to multimorbidity associated with myocardial infarction."
  BMC Med Res Methodol. 25(1):32. <doi:10.1186/s12874-025-02476-7>

Maintainer: Marc Delord <mdelord@gmail.com>

}

\keyword{ clustering }
