% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f2.R
\name{msca}
\alias{msca}
\title{Compute Jaccard Index Using Parallel Rcpp}
\usage{
msca(smats)
}
\arguments{
\item{smats}{A numeric matrix where Jaccard Index will be computed between the columns.}
}
\value{
A matrix containing the Jaccard index between columns.
}
\description{
This function takes a matrix, removes rows where all values are zero (or NA),
and computes the Jaccard index using an Rcpp parallel function.
}
